update： - SSE 推送新增系统组装的完整提示词内容（含拉取封装的virus total数据） - 页面差异化展示模型思考内容（后端非标准方式提取）与正文内容 - 前端对提示词和思考内容增加折叠/展开功能 - 优化页面在 SSE 接收状态下的滚动平滑度，保障流畅滚动 - 调整 DnsRecord 实体，将 serial 属性由 Integer 调整为 Long 提升溢出防护能力

+++ src/main/java/com/vt/remote/dto/vt/domain/sub/DnsRecord.java
@@ -19,7 +19,7 @@ public record DnsRecord(
         @Schema(description = "刷新间隔 (Refresh Interval)。辅 DNS 服务器检查主 DNS 服务器序列号以获取更新的时间间隔")
         Long retry,
         @Schema(description = "序列号 (Serial Number)。区域文件的版本号，当区域数据更新时，此值必须递增，以便辅 DNS 服务器判断是否需要同步")
-        Integer serial,
+        Long serial,
         @Schema(description = "通常与 type: \"CAA\" (证书颁发机构授权) 记录绑定，用于指定要限制的属性，例如 issue (授权颁发) 或 issuewild (授权通配符证书)")
         String tag,
         @Schema(description = "规定该条 DNS 记录在 DNS 解析器上的缓存时间长度，单位为秒。超过此时间，解析器会重新查询")
